+++ shared_model/backend/protobuf/query_responses/proto_engine_response_record.hpp
@@ -6,7 +6,7 @@
 #ifndef IROHA_SHARED_PROTO_MODEL_ENGINE_RECEIPT_HPP
 #define IROHA_SHARED_PROTO_MODEL_ENGINE_RECEIPT_HPP
 
-#include "interfaces/query_responses/engine_response_record.hpp"
+#include "interfaces/query_responses/engine_receipt.hpp"
 
 #include "cryptography/hash.hpp"
 #include "interfaces/common_objects/types.hpp"
