+++ shared_model/interfaces/query_responses/engine_receipt.hpp
