+++ shared_model/interfaces/query_responses/impl/engine_receipt.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "interfaces/query_responses/engine_response_record.hpp"
+#include "interfaces/query_responses/engine_receipt.hpp"
 
 #include <iostream>
