+++ test/module/shared_model/interface_mocks.hpp
@@ -15,7 +15,7 @@
 #include "interfaces/iroha_internal/proposal.hpp"
 #include "interfaces/iroha_internal/transaction_batch.hpp"
 #include "interfaces/iroha_internal/unsafe_proposal_factory.hpp"
-#include "interfaces/query_responses/engine_response_record.hpp"
+#include "interfaces/query_responses/engine_receipt.hpp"
 #include "interfaces/transaction.hpp"
 
 // TODO: 2019-01-18 @muratovv Separate file by classes IR-229
